requirements.txt
aiogram
pyrebase4